pyproject.toml
[tool.poetry]
name = "adventOfCode"
version = "0.1.0"
description = ""
authors = ["Yrieix Leprince <y.leprince@criteo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
